+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-06-25 20:49:31
-// Commit: feat(core): integrate custom telemetry logger in core
+// Updated at: 2026-06-25 22:00:26
+// Commit: perf(core): apply RepaintBoundary on animated core widgets
 
 class coreComponentState {
-  final String commitRef = '2330';
-  final String lastMessage = 'feat(core): integrate custom telemetry logger in core';
-  final DateTime timestamp = DateTime.parse('2026-06-25T20:49:31+05:00');
+  final String commitRef = '2339';
+  final String lastMessage = 'perf(core): apply RepaintBoundary on animated core widgets';
+  final DateTime timestamp = DateTime.parse('2026-06-25T22:00:26+05:00');
 }